move GCSE_SECTIONS to constants and replace magic strings with named constants
